- Do not add equations if we bind a record variable using a ceval'ed record constructor. - Do not change variability of record fields from const to param if the parent is declared param.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@4493 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/java_interface.c
@@ -683,7 +683,7 @@ jobject mmc_to_jobject(JNIEnv* env, void* mmc)
   unsigned ctor;
   int i;
 
-  if (0 == ((int)mmc & 1)) /* INTEGER */
+  if (0 == ((long)mmc & 1)) /* INTEGER */
     return NewJavaInteger(env,MMC_UNTAGFIXNUM(mmc));
   hdr = MMC_GETHDR(mmc);
   if (hdr == MMC_REALHDR) /* REAL */
